requirements.txt
GitPython
chromadb
torch
tqdm
transformers
radon
